remove dependency of geometric_elements to AdelR
